Package up base db and atomic db test suites into re-usable classes (#1813)

+++ eth/db/backends/level.py
@@ -62,6 +62,8 @@ class LevelDB(BaseAtomicDB):
         return self.db.get(key) is not None
 
     def __delitem__(self, key: bytes) -> None:
+        if self.db.get(key) is None:
+            raise KeyError(key)
         self.db.delete(key)
 
     @contextmanager
